# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /search:
    get:
      operationId: searchContent
      tags: [search]
      summary: Search across video frames and transcriptions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [all, visual, audio]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
...
  /search/feedback:
    post:
      operationId: submitSearchFeedback
      tags: [search]
      summary: Submit thumbs up/down feedback on a search result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchFeedbackBody"
      responses:
        "200":
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SearchResults:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        total:
          type: integer
        query:
          type: string
      required:
        - results
        - total
        - query
    SearchResult:
      type: object
      properties:
        type:
          type: string
          enum: [frame, transcription]
        videoId:
          type: integer
        videoTitle:
          type: string
        driveFileId:
          type: ["string", "null"]
        timestampSec:
          type: number
        endSec:
          type: ["number", "null"]
        content:
          type: string
        imagePath:
          type: ["string", "null"]
        allFramePaths:
          type: array
          items:
            type: string
        rank:
          type: number
        videoTags:
          type: ["string", "null"]
      required:
        - type
        - videoId
        - videoTitle
        - timestampSec
        - content
        - rank
...
    SearchFeedbackBody:
      type: object
      properties:
        videoId:
          type: integer
        query:
          type: string
        feedbackType:
          type: string
          enum: [up, down]
      required:
        - videoId
        - query
        - feedbackType
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...