# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResultType } from "./searchResultType";

export interface SearchResult {
  type: SearchResultType;
  videoId: number;
  videoTitle: string;
  /** @nullable */
  driveFileId?: string | null;
  timestampSec: number;
  /** @nullable */
  endSec?: number | null;
  content: string;
  /** @nullable */
  imagePath?: string | null;
  allFramePaths?: string[];
  rank: number;
  /** @nullable */
  videoTags?: string | null;
}